requirements.txt
pandas==1.1.3
streamlit==1.3.0
regex==2020.10.15
mysql-connector-python==8.0.27
plotly-express==0.4.0
nltk==3.5
numpy==1.21.2
matplotlib==3.3.2
wordcloud==1.8.1
tweepy==4.4.0
textblob==0.17.1